Realign volumes for SaaS/PaaS

+++ scripts/scaffold-existing-project-add-lando.sh
@@ -12,6 +12,7 @@ NC='\033[0m' # No Color
 ################## TODO "my-project" is "my" instead?
 GOVCMS_NAME=$(grep -oP 'http:\/\/\K[^\n]+(?=\.docker\.amazee\.io)' docker-compose.yml | head -n 1)
 GOVCMS_VERSION=$(grep -oP 'GOVCMS_IMAGE_VERSION:-\K\w+' docker-compose.yml | head -n 1)
+GOVCMS_TYPE=$(grep -oP '^type:\s+\K\w+' .version.yml | head -n 1)
 grep -q '^  solr:$' docker-compose.yml
 HAS_SOLR=$?
 
@@ -25,12 +26,23 @@ if [[ -z "$GOVCMS_VERSION" ]]; then
   exit 2
 fi
 
+if [[ -z "$GOVCMS_TYPE" ]]; then
+  echo "[error]: Cannot determine scaffold type, must be (saas, saasplus or paas)."
+  exit 2
+fi
+
 echo "[info]: Modifying scaffold for GovCMS$GOVCMS_VERSION: $GOVCMS_NAME"
 
 sed -i.bak "s/{{ GOVCMS_PROJECT_NAME }}/$GOVCMS_NAME/" .lando.base.yml && rm .lando.base.yml.bak
 sed -i.bak "s/{{ GOVCMS_PROJECT_NAME }}/$GOVCMS_NAME/" .lando.local.example.yml && rm .lando.local.example.yml.bak
 sed -i.bak "s/{{ GOVCMS_VERSION }}/$GOVCMS_VERSION/" .lando.base.yml && rm .lando.base.yml.bak
 sed -i.bak "s/{{ GOVCMS_VERSION }}/$GOVCMS_VERSION/" .lando.local.example.yml && rm .lando.local.example.yml.bak
+
+if [[ "$GOVCMS_TYPE" != "paas" ]]; then
+  : # Do nothing.
+else
+  sed -i.bak -E 's/(\*default-volumes)([[:space:]]|$|#)/\*paas-volumes\2/' .lando.base.yml && rm .lando.base.yml.bak
+fi
 
 if [[ "$HAS_SOLR" -eq 0 ]]; then
   printf "\n${YELLOW}[!! action required !!]${NC}: SOLR has been detected,\n"
